+++ chat_api/models/message.py
@@ -5,7 +5,7 @@ from ..extensions import db
 class Message(db.Model):
     __tablename__ = "messages"
 
-    message_id = db.Column(db.Integer, primary_key=True)
+    message_id = db.Column(db.BigInteger, primary_key=True)
 
     chat_id = db.Column(
         db.Integer,
